blog/requirements.txt
asgiref==3.7.2
blinker==1.6.2
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
command-not-found==0.3
crispy-bootstrap4==2023.1
crispy-bootstrap5==2023.10
cryptography==3.4.8
dbus-python==1.2.18
distlib==0.3.7
distro==1.7.0
distro-info==1.1+ubuntu0.1
Django==4.2.5
django-crispy-forms==2.1
djangorestframework==3.14.0
filelock==3.12.4
Flask==2.3.3
Flask-Cors==4.0.0
greenlet==2.0.2
httplib2==0.20.2
idna==3.4
importlib-metadata==4.6.4
itsdangerous==2.1.2
jeepney==0.7.1
Jinja2==3.1.2
keyring==23.5.0
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
MarkupSafe==2.1.3
more-itertools==8.10.0
mysqlclient==2.1.1
netifaces==0.11.0
oauthlib==3.2.0
pep8==1.7.1
Pillow==10.1.0
pipenv==11.9.0
platformdirs==3.10.0
pycodestyle==2.8.0
PyGObject==3.42.1
PyJWT==2.3.0
pyparsing==2.4.7
python-apt==2.4.0+ubuntu2
pytz==2023.3.post1
PyYAML==5.4.1
requests==2.31.0
SecretStorage==3.3.1
six==1.16.0
SQLAlchemy==2.0.20
sqlparse==0.4.4
systemd-python==234
typing_extensions==4.7.1
ubuntu-advantage-tools==8001
ufw==0.36.1
unattended-upgrades==0.1
urllib3==2.0.5
virtualenv==20.24.5
virtualenv-clone==0.3.0
wadllib==1.3.6
Werkzeug==2.3.7
zipp==1.0.0
